+++ tests/helper.rs
@@ -24,7 +24,7 @@ macro_rules! help_msg {
             Option Parameters:
               <when>    'always', 'never', or 'auto'
               <exp>     regular expression can has capture groups
-              <fmt>     format can has capture group: $0, $1, $2, ...
+              <fmt>     format can has capture group: ${0}, ${1}, ${2}, ...
 
             Environments:
               AKI_GSUB_COLOR_SEQ_ST     color start sequence specified by ansi
@@ -33,7 +33,7 @@ macro_rules! help_msg {
             Examples:
               Leaving one character between 'a' and 'c', converts 'a' and 'c'
               on both sides to '*':
-                echo "abcabca" | aki-gsub -e "a(.)c" -f "*\$1*"
+                echo "abcabca" | aki-gsub -e "a(.)c" -f "*\${1}*"
               result output:
                 *b**b*a